UPSTREAM: fuse: add readdir cache version
Allow the cache to be invalidated when page(s) have gone missing.  In this
case increment the version of the cache and reset to an empty state.

Add a version number to the directory stream in struct fuse_file as well,
indicating the version of the cache it's supposed to be reading.  If the
cache version doesn't match the stream's version, then reset the stream to
the beginning of the cache.

Change-Id: I2458cc0c62dcea26f8462dd243fe72be0d4b76a6
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>