UPSTREAM: fuse: add simple background helper
Create a helper named fuse_simple_background() that is similar to
fuse_simple_request().  Unlike the latter, it returns immediately and calls
the supplied 'end' callback when the reply is received.

The supplied 'args' pointer is stored in 'fuse_req' which allows the
callback to interpret the output arguments decoded from the reply.

Change-Id: I9b2626d601d3b1508bbcc0ef99fc41203aeb2f8a
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>